Move cosmosmemoryrecord out to it's own class, remove a bunch of code as the kernel made it public
