--- dotnet/src/SemanticKernel.Skills/Skills.Memory.CosmosDB/Collections/TopNCollection.cs
@@ -1,102 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-
-using System.Collections;
-using System.Collections.Generic;
-
-namespace Microsoft.SemanticKernel.Memory.Collections;
-
-/// <summary>
-/// A collector for Top N matches. Keeps only the best N matches by Score.
-/// Automatically flushes out any not in the top N.
-/// By default, items are not sorted by score until you call <see cref="TopNCollection{T}.SortByScore"/>.
-/// </summary>
-internal class TopNCollection<T> : IEnumerable<ScoredValue<T>>
-{
-    private readonly int _maxItems;
-    private readonly MinHeap<ScoredValue<T>> _heap;
-    private bool _sorted = false;
-
-    public TopNCollection(int maxItems)
-    {
-        this._maxItems = maxItems;
-        this._heap = new MinHeap<ScoredValue<T>>(ScoredValue<T>.Min(), maxItems);
-    }
-
-    public int MaxItems => this._maxItems;
-    public int Count => this._heap.Count;
-
-    internal ScoredValue<T> this[int i] => this._heap[i];
-    internal ScoredValue<T> Top => this._heap.Top;
-
-    /// <summary>
-    /// Call this to reuse the buffer
-    /// </summary>
-    public void Reset()
-    {
-        this._heap.Clear();
-    }
-
-    /// <summary>
-    /// Adds a single scored value
-    /// </summary>
-    public void Add(ScoredValue<T> value)
-    {
-        if (this._sorted)
-        {
-            this._heap.Restore();
-            this._sorted = false;
-        }
-
-        if (this._heap.Count == this._maxItems)
-        {
-            // Queue is full. We will need to dequeue the item with lowest weight
-            if (value.Score <= this.Top.Score)
-            {
-                // This score is lower than the lowest score on the queue right now. Ignore it
-                return;
-            }
-
-            this._heap.RemoveTop();
-        }
-
-        this._heap.Add(value);
-    }
-
-    public void Add(T value, Score score)
-    {
-        this.Add(new ScoredValue<T>(value, score));
-    }
-
-    /// <summary>
-    /// Sort in relevancy order.
-    /// </summary>
-    public void SortByScore()
-    {
-        if (!this._sorted && this._heap.Count > 0)
-        {
-            this._heap.SortDescending();
-            this._sorted = true;
-        }
-    }
-
-    public IList<ScoredValue<T>> ToList()
-    {
-        var list = new List<ScoredValue<T>>(this.Count);
-        for (int i = 0, count = this.Count; i < count; ++i)
-        {
-            list.Add(this[i]);
-        }
-
-        return list;
-    }
-
-    public IEnumerator<ScoredValue<T>> GetEnumerator()
-    {
-        return this._heap.GetEnumerator();
-    }
-
-    IEnumerator IEnumerable.GetEnumerator()
-    {
-        return this._heap.GetEnumerator();
-    }
-}
